requirements.txt
streamlit>=1.44.0
python-binance>=1.0.29
pandas>=2.2.0
python-dotenv>=1.0.1
